+++ hydromt_wflow/workflows/__init__.py
@@ -2,10 +2,13 @@
 
 from .basemaps import *
 from .connect import *
+from .forcing import *
 from .gauges import *
 from .glaciers import *
 from .landuse import *
 from .river import *
 from .rootzoneclim import *
 from .soilgrids import *
+from .soilparams import *
+from .states import *
 from .waterbodies import *