+++ hydromt_wflow/workflows/soilgrids.py
@@ -1,7 +1,9 @@
 """Soilgrid workflows for Wflow plugin."""
 
 import logging
+from typing import List
 
+import hydromt
 import numpy as np
 import pandas as pd
 import xarray as xr
@@ -11,7 +13,7 @@ from . import ptf
 
 logger = logging.getLogger(__name__)
 
-__all__ = ["soilgrids", "soilgrids_sediment"]
+__all__ = ["soilgrids", "soilgrids_sediment", "soilgrids_brooks_corey"]
 
 # soilgrids_2017
 soildepth_cm = np.array([0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0])
@@ -28,6 +30,46 @@ nodata = -9999.0
 c_sl_index = [2, 4, 6, 7]  # v2017 direct mapping
 # c_sl_index = [2, 3, 5, 6] #v2021 if direct mapping - not used,
 # averages are taken instead.
+
+
+def concat_layers(
+    ds: xr.Dataset,
+    soil_fn: str = "soilgrids",
+    variables: List[str] = ["bd", "oc", "ph", "clyppt", "sltppt", "sndppt"],
+):
+    """
+    Preprocess functions to concat soilgrids along a layer dimension.
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+        Dataset containing soil properties.
+    soil_fn : str
+        soilgrids version {'soilgrids', 'soilgrids_2020'}
+    variables : list
+        List of soil properties to concat.
+    """
+    if soil_fn == "soilgrids_2020":
+        nb_sl = 6
+    else:
+        nb_sl = 7
+    ds = ds.assign_coords(sl=np.arange(1, nb_sl + 1))
+
+    for var in variables:
+        da_prop = []
+        for i in np.arange(1, nb_sl + 1):
+            da_prop.append(ds[f"{var}_sl{i}"])
+            # remove layer from ds
+            ds = ds.drop_vars(f"{var}_sl{i}")
+        da = xr.concat(
+            da_prop,
+            pd.Index(np.arange(1, nb_sl + 1, dtype=int), name="sl"),
+        ).transpose("sl", ...)
+        da.name = var
+        # add concat maps to ds
+        ds[f"{var}"] = da
+
+    return ds
 
 
 def average_soillayers_block(ds, soilthickness):
@@ -165,6 +207,121 @@ def pore_size_distrution_index_layers(ds, thetas):
     return ds_out
 
 
+def brooks_corey_layers(
+    thetas_sl: xr.Dataset,
+    ds: xr.Dataset,
+    ds_like: xr.Dataset,
+    soil_fn: str = "soilgrids",
+    wflow_layers: List[int] = [100, 300, 800],
+    soildepth_cm: np.array = np.array([0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0]),
+):
+    """
+    Determine Brooks Corey coefficient per wflow soil layer depth.
+
+    First pore size distribution index is computed based on theta_s and other soil
+    parameters and scaled to the model resolution.
+
+    Then the Brooks Corey coefficient is computed for each wflow soil layer depth
+    by weighted averaging the pore size distribution index over the soil thickness.
+
+    Parameters
+    ----------
+    thetas_sl: xarray.Dataset
+        Dataset containing thetaS at each soil layer depth.
+    ds: xarray.Dataset
+        Dataset containing soil properties at each soil depth.
+    ds_like: xarray.Dataset
+        Dataset at model resolution for reprojection.
+    soil_fn: str
+        soilgrids version {'soilgrids', 'soilgrids_2020'}
+    wflow_layers: list
+        List of soil layer depths [cm] for which c is calculated.
+    soildepth_cm: np.array
+        Depth of each soil layers [cm].
+
+    Returns
+    -------
+    ds_c : xarray.Dataset
+        Dataset containing c for the wflow_sbm soil layers.
+    """
+    # Get pore size distribution index
+    lambda_sl_hr = pore_size_distrution_index_layers(ds, thetas_sl)
+    lambda_sl = np.log(lambda_sl_hr)
+    lambda_sl = lambda_sl.raster.reproject_like(ds_like, method="average")
+    lambda_sl = np.exp(lambda_sl)
+
+    # Brooks Corey coefficient
+    c_sl = 3.0 + (2.0 / lambda_sl)
+    c_sl.name = "c_sl"
+
+    # Resample for the wflow layers
+    wflow_thickness = wflow_layers.copy()
+    # Go from wflow layer thickness to soil depths (cumulative)
+    wflow_depths = np.cumsum(wflow_thickness)
+    # Check if the last wflow depth is less than 2000 mm (soilgrids limit)
+    if wflow_depths[-1] > 2000:
+        raise ValueError(
+            "The total depth of the wflow soil layers should be 2000 mm, as the \
+soilgrids data does not go deeper than 2 m."
+        )
+    # Add a zero for the first depth and 2000 for the last depth
+    wflow_depths = np.insert(wflow_depths, 0, 0)
+    wflow_depths = np.append(wflow_depths, 2000)
+    # Compute the thickness of the last layer
+    wflow_thickness.append(2000 - wflow_depths[-2])
+
+    # Soil data depth
+    soildepth = soildepth_cm * 10
+
+    # make empty dataarray for c for the 4 sbm layers
+    ds_c = hydromt.raster.full(
+        coords=c_sl.raster.coords,
+        nodata=np.nan,
+        crs=c_sl.raster.crs,
+        dtype="float32",
+        name="c",
+        attrs={"thickness": wflow_thickness, "soil_fn": soil_fn},
+    )
+    ds_c = ds_c.expand_dims(dim=dict(layer=np.arange(len(wflow_thickness)))).copy()
+
+    # calc weighted average values of c over the sbm soil layers
+    for nl in range(len(wflow_thickness)):
+        top_depth = wflow_depths[nl]
+        bottom_depth = wflow_depths[nl + 1]
+        c_nl_sum = None
+        for d in range(len(soildepth) - 1):
+            if soil_fn == "soilgrids_2020":
+                c_av = c_sl.sel(sl=d + 1)
+            else:
+                c_av = (c_sl.sel(sl=d + 1) + c_sl.sel(sl=d + 2)) / 2
+            # wflow layer fully within soilgrid layer
+            if soildepth[d] <= top_depth and soildepth[d + 1] >= bottom_depth:
+                c_nl = c_av * (bottom_depth - top_depth)
+            # layer fully within wflow layer
+            elif soildepth[d] >= top_depth and soildepth[d + 1] <= bottom_depth:
+                c_nl = c_av * (soildepth[d + 1] - soildepth[d])
+            # bottom part of the layer wihtin wflow
+            elif soildepth[d] <= bottom_depth and soildepth[d + 1] >= bottom_depth:
+                c_nl = c_av * (bottom_depth - soildepth[d])
+            # top part of the layer within wflow
+            elif soildepth[d] <= top_depth and soildepth[d + 1] >= top_depth:
+                c_nl = c_av * (soildepth[d + 1] - top_depth)
+            # layer outside of wflow layer
+            else:
+                c_nl = None
+
+            # Add to the sum
+            if c_nl is not None:
+                if c_nl_sum is None:
+                    c_nl_sum = c_nl
+                else:
+                    c_nl_sum = c_nl_sum + c_nl
+
+        ds_c.loc[dict(layer=nl)] = c_nl_sum / wflow_thickness[nl]
+
+    return ds_c
+
+
 def kv_layers(ds, thetas, ptf_name):
     """
     Determine vertical saturated hydraulic conductivity (KsatVer) per soil layer depth.
@@ -283,7 +440,14 @@ def constrain_M(M, popt_0, M_minmax):
     return M
 
 
-def soilgrids(ds, ds_like, ptfKsatVer, soil_fn, logger=logger):
+def soilgrids(
+    ds: xr.Dataset,
+    ds_like: xr.Dataset,
+    ptfKsatVer: str = "brakensiek",
+    soil_fn: str = "soilgrids",
+    wflow_layers: List[int] = [100, 300, 800],
+    logger=logger,
+):
     """
     Return soil parameter maps at model resolution.
 
@@ -319,12 +483,8 @@ soil depth
 (fitted with curve_fit (scipy.optimize)), bounds are checked
     - **f_** : scaling parameter controlling the decline of KsatVer [mm-1]
       (fitted with numpy linalg regression), bounds are checked
-    - **c_0** : Brooks Corey coefficient [-] based on pore size distribution index at \
-depth 1st
-      soil layer (100 mm) wflow_sbm
-    - **c_1** : idem **c_0** at depth 2nd soil layer (400 mm) wflow_sbm
-    - **c_2** : idem **c_0** at depth 3rd soil layer (1200 mm) wflow_sbm
-    - **c_3** : idem **c_0** at depth 4th soil layer (> 1200 mm) wflow_sbm
+    - **c_** map: Brooks Corey coefficients [-] based on pore size distribution \
+index for the wflow_sbm soil layers.
     - **KsatVer_[z]cm** : KsatVer [mm/day] at soil depths [z] of SoilGrids data \
 [0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0]
     - **wflow_soil** : USDA Soil texture based on percentage clay, silt, sand mapping: \
@@ -342,6 +502,8 @@ depth 1st
         PTF to use for calculcation KsatVer.
     soil_fn : str
         soilgrids version {'soilgrids', 'soilgrids_2020'}
+    wflow_layers : list
+        List of soil layer depths [cm] for which c is calculated.
 
     Returns
     -------
@@ -353,14 +515,10 @@ depth 1st
         soildepth_cm_midpoint = np.array([2.5, 10.0, 22.5, 45.0, 80.0, 150.0])
         soildepth_cm_midpoint_surface = np.array([0, 10.0, 22.5, 45.0, 80.0, 150.0])
         soildepth_cm = np.array([0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0])
-        10.0 * soildepth_cm
-        10.0 * soildepth_cm_midpoint
         soildepth_mm_midpoint_surface = 10.0 * soildepth_cm_midpoint_surface
     else:
         soildepth_cm = np.array([0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0])
-        10.0 * soildepth_cm
         soildepth_cm_midpoint = np.array([0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0])
-        10.0 * soildepth_cm
         soildepth_mm_midpoint_surface = 10.0 * soildepth_cm
 
     ds_out = xr.Dataset(coords=ds_like.raster.coords)
@@ -370,24 +528,8 @@ depth 1st
     # ds = xr.where(ds["sltppt_sl1"] == ds["sltppt_sl1"].raster.nodata, np.nan, ds)
     ds = ds.raster.mask_nodata()
 
-    # add new coordinate sl to merge layers
-    ds = ds.assign_coords(sl=np.arange(1, len(soildepth_cm_midpoint) + 1))
-
-    for var in ["bd", "oc", "ph", "clyppt", "sltppt", "sndppt"]:
-        da_prop = []
-        for i in np.arange(1, len(soildepth_cm_midpoint) + 1):
-            da_prop.append(ds[f"{var}_sl{i}"])
-            # remove layer from ds
-            ds = ds.drop_vars(f"{var}_sl{i}")
-        da = xr.concat(
-            da_prop,
-            pd.Index(
-                np.arange(1, len(soildepth_cm_midpoint) + 1, dtype=int), name="sl"
-            ),
-        ).transpose("sl", ...)
-        da.name = var
-        # add concat maps to ds
-        ds[f"{var}"] = da
+    # concat along a sl dimension
+    ds = concat_layers(ds, soil_fn)
 
     logger.info("calculate and resample thetaS")
     thetas_sl = xr.apply_ufunc(
@@ -442,40 +584,14 @@ depth 1st
     kv_sl = np.exp(kv_sl)
 
     logger.info("calculate and resample pore size distribution index")
-    lambda_sl_hr = pore_size_distrution_index_layers(ds, thetas_sl)
-    lambda_sl = np.log(lambda_sl_hr)
-    lambda_sl = lambda_sl.raster.reproject_like(ds_like, method="average")
-    lambda_sl = np.exp(lambda_sl)
-
-    if soil_fn == "soilgrids_2020":
-        # calculate c for each layer of lambda
-        c_sl = 3.0 + (2.0 / lambda_sl)
-        c_sl.name = "c_sl"
-
-        # make empty dataarray for c for the 4 sbm layers
-        ds_c = xr.full_like(soilthickness, np.nan)
-        ds_c.name = "c"
-        ds_c = ds_c.expand_dims(dim=dict(layer=[0, 1, 2, 3])).copy()
-        # calc weighted average values of c over the sbm soil layers
-        ds_c.loc[dict(layer=0)] = (
-            c_sl.sel(sl=1) * 50 + c_sl.sel(sl=2) * 50
-        ) / 100  # average over layer 0-100
-        ds_c.loc[dict(layer=1)] = (
-            c_sl.sel(sl=2) * 50 + c_sl.sel(sl=3) * 150 + c_sl.sel(sl=4) * 100
-        ) / 300  # average over layer 100-400
-        ds_c.loc[dict(layer=2)] = (
-            c_sl.sel(sl=4) * 200 + c_sl.sel(sl=5) * 400 + c_sl.sel(sl=6) * 200
-        ) / 800  # average over layer 400-1200
-        ds_c.loc[dict(layer=3)] = c_sl.sel(sl=6)  # layer >1200
-    else:
-        # for soilgrids 2017, keep the direct mapping calculation.
-        da_c = []
-        for i, sl_ind in enumerate(c_sl_index):
-            da_c.append(3.0 + (2.0 / lambda_sl.sel(sl=sl_ind)))
-        ds_c = xr.concat(
-            da_c, pd.Index(np.arange(len(c_sl_index), dtype=int), name="layer")
-        ).transpose("layer", ...)
-        ds_c.name = "c"
+    ds_c = brooks_corey_layers(
+        thetas_sl=thetas_sl,
+        ds=ds,
+        ds_like=ds_like,
+        soil_fn=soil_fn,
+        wflow_layers=wflow_layers,
+        soildepth_cm=soildepth_cm,
+    )
     ds_out["c"] = ds_c
 
     ds_out["KsatVer"] = kv_sl.sel(sl=1).astype(np.float32)
@@ -561,6 +677,75 @@ depth 1st
         ds_out[var] = ds_out[var].raster.interpolate_na("nearest")
         ds_out[var] = ds_out[var].fillna(nodata).astype(dtype)
         ds_out[var].raster.set_nodata(np.dtype(dtype).type(nodata))
+
+    return ds_out
+
+
+def soilgrids_brooks_corey(
+    ds: xr.Dataset,
+    ds_like: xr.Dataset,
+    soil_fn: str = "soilgrids",
+    wflow_layers: List[int] = [100, 300, 800],
+    logger=logger,
+):
+    """
+    Determine Brooks Corey coefficient per wflow soil layer depth.
+
+    First pore size distribution index is computed based on theta_s and other soil
+    parameters and scaled to the model resolution.
+
+    Then the Brooks Corey coefficient is computed for each wflow soil layer depth
+    by weighted averaging the pore size distribution index over the soil thickness.
+
+    Parameters
+    ----------
+    ds : xarray.Dataset
+        Dataset containing soil properties.
+    ds_like : xarray.DataArray
+        Dataset at model resolution.
+    soil_fn : str
+        soilgrids version {'soilgrids', 'soilgrids_2020'}
+    wflow_layers : list
+        List of wflow soil layer depths [cm] over which c is calculated.
+
+    Returns
+    -------
+    ds_out : xarray.Dataset
+        Dataset containing c for the wflow_sbm soil layers.
+    """
+    if soil_fn == "soilgrids_2020" or soil_fn == "soilgrids":
+        soildepth_cm = np.array([0.0, 5.0, 15.0, 30.0, 60.0, 100.0, 200.0])
+    else:
+        raise ValueError("Only soilgrids_2020 and soilgrids are supported.")
+
+    ds_out = xr.Dataset(coords=ds_like.raster.coords)
+    ds = ds.raster.mask_nodata()
+    # concat along a sl dimension
+    ds = concat_layers(ds, soil_fn)
+
+    logger.info("calculate and resample thetaS")
+    thetas_sl = xr.apply_ufunc(
+        ptf.thetas_toth,
+        ds["ph"],
+        ds["bd"],
+        ds["clyppt"],
+        ds["sltppt"],
+        dask="parallelized",
+        output_dtypes=[float],
+        keep_attrs=True,
+    )
+
+    logger.info("calculate and resample pore size distribution index")
+    # Brooks Corey coefficient
+    ds_c = brooks_corey_layers(
+        thetas_sl=thetas_sl,
+        ds=ds,
+        ds_like=ds_like,
+        soil_fn=soil_fn,
+        wflow_layers=wflow_layers,
+        soildepth_cm=soildepth_cm,
+    )
+    ds_out["c"] = ds_c
 
     return ds_out