+++ docs/changelog.rst
@@ -11,16 +11,24 @@ Unreleased
 
 Added
 -----
+- New function **setup_1dmodel_connection** to connect wflow to 1D river model (eg Delft3D FM 1D, HEC-RAS, etc.) `PR #210 <https://github.com/Deltares/hydromt_wflow/pull/210>`_
+- New setup method for the **KsatHorFrac** parameter **setup_ksathorfarc** to up-downscale existing ksathorfrac maps. `PR #255 <https://github.com/Deltares/hydromt_wflow/pull/255>`_
+- new function **setup_pet_forcing** to reproject existing pet data rather than computing from other meteo data. PR #257
+- Workflow to compute brooks corey c for the wflow layers based on soilgrids data, soilgrids_brooks_corey. PR #242
 - **setup_lulcmaps** for wflow_sediment: if planted forest data is available, it can be used to update the values of the USLE C parameter. PR #234
-- new function **setup_1dmodel_connection** to connect wflow to 1D river model (eg Delft3D FM 1D, HEC-RAS, etc.) `PR #210 <https://github.com/Deltares/hydromt_wflow/pull/210>`_
+- better support for WflowModel states with new methods: **read_states**, **write_states** and **clip_states**. PR #252
+- new function **setup_cold_states** to prepare cold states for WflowModel. PR #252
+- new utils method **get_grid_from_config** to get the right wflow staticmaps variable based on the TOML configuration (e.g. detects name in netcdf, value, scale and offset). Only applied now to prepare cold states (e.g. not yet in read_grid). PR #252
 
 Changed
 -------
+- **setup_soilmaps**: the user can now supply variable sbm soil layer thicknesses. The Brooks Corey coefficient is then computed as weighted average over the sbm layers. PR #242
 - **setup_outlets**: the IDs of wflow_subcatch are used to define the outlets IDs rather than [1:n]. PR #247
 
 Fixed
 -----
-- Wrong dtype for wflow_subcatch map. PR #247
+- Wrong dtype for wflow_subcatch map. PR #247,
+- Removed building a wflow model without a config file in the build notebook.
 
 v0.5.0 (February 2024)
 ======================